coding/middleware/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0